requirements.txt
flask
pandas
numpy
google-generativeai
sentence-transformers
faiss-cpu
transformers
torch
jinja2
faker